Moved authenticated data into the same socket connection as the rest of the
