+++ service-netty/src/main/java/info/bitrich/xchangestream/service/netty/JsonNettyStreamingService.java
@@ -1,8 +1,10 @@
 package info.bitrich.xchangestream.service.netty;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.JsonNodeType;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -40,6 +42,14 @@ public abstract class JsonNettyStreamingService extends NettyStreamingService<Js
             }
         } else {
             handleMessage(jsonNode);
+        }
+    }
+
+    protected void sendObjectMessage(Object message) {
+        try {
+            sendMessage(objectMapper.writeValueAsString(message));
+        } catch (JsonProcessingException e) {
+            LOG.error("Error creating json message: {}", e.getMessage());
         }
     }
 }
